refactored organization components
